Move tests alongside source files, organise fixtures
