+++ test/form/rate-limit.test.js
@@ -1,4 +1,4 @@
-import { dirname } from 'node:path'
+import { dirname, join } from 'node:path'
 import { fileURLToPath } from 'node:url'
 
 import { createServer } from '~/src/server/index.js'
@@ -20,7 +20,7 @@ describe('Rate limit', () => {
   beforeAll(async () => {
     server = await createServer({
       formFileName: 'basic-v1.json',
-      formFilePath: testDir,
+      formFilePath: join(testDir, 'definitions'),
       rateOptions: options
     })
     server.route({
